+++ admin/js.js
@@ -1,8 +1,8 @@
 jQuery(document).ready(function($) {
     /**
-     * Add rating hint.
+     * Add rating hint and add file action.
      */
-    $('body.settings_page_eml_settings h1.wp-heading-inline, body.media_page_efml_local_directories h1.wp-heading-inline, body.taxonomy-edlfw_archive h1.wp-heading-inline, body.media_page_efml_local_directories h1.wp-heading-inline').each(function() {
+    $('body.settings_page_eml_settings h1.wp-heading-inline, body.taxonomy-edlfw_archive h1.wp-heading-inline').each(function() {
       let review_button = document.createElement('a');
       review_button.className = 'review-hint-button page-title-action';
       review_button.href = efmlJsVars.review_url;
@@ -15,17 +15,24 @@ jQuery(document).ready(function($) {
       add_file_button.href = efmlJsVars.add_file_url;
       add_file_button.innerHTML = efmlJsVars.title_add_file;
       this.after(add_file_button);
-    });
 
-    /**
-     * Add link to overview over possible external sources.
-     */
-    $('body.taxonomy-edlfw_archive h1.wp-heading-inline').each(function() {
       let add_directory_listing_button = document.createElement( 'a' );
       add_directory_listing_button.className = 'page-title-action';
       add_directory_listing_button.href = efmlJsVars.directory_listing_url;
       add_directory_listing_button.innerHTML = efmlJsVars.title_add_source;
       this.after( add_directory_listing_button );
+    });
+
+    /**
+     * Add rating hint.
+     */
+    $('body.media_page_efml_local_directories h1.wp-heading-inline').each(function() {
+      let review_button = document.createElement('a');
+      review_button.className = 'review-hint-button page-title-action';
+      review_button.href = efmlJsVars.review_url;
+      review_button.innerHTML = efmlJsVars.title_rate_us;
+      review_button.target = '_blank';
+      this.after(review_button);
     });
 
     /**
@@ -92,7 +99,7 @@ jQuery(document).ready(function($) {
         let id = $("#post_ID").val();
 
         // send request
-        jQuery.ajax({
+        $.ajax({
           url: efmlJsVars.ajax_url,
           type: 'post',
           data: {
@@ -264,15 +271,19 @@ function efml_upload_files() {
   let additional_fields = {};
   jQuery('.easy-dialog-for-wordpress-text .eml-use-for-import').each(function() {
     if( 'INPUT' === jQuery(this).prop('nodeName') ) {
-      if( 'checkbox' === jQuery(this).attr('type') && jQuery(this).prop('checked') === true ) {
-        if( jQuery(this).hasClass('eml-multi') ) {
-          if (!additional_fields[jQuery( this ).prop( 'name' )]) {
-            additional_fields[jQuery( this ).prop( 'name' )] = {};
+      if( 'checkbox' === jQuery(this).attr('type') ) {
+        if( jQuery(this).prop('checked') === true ) {
+          if (jQuery( this ).hasClass( 'eml-multi' )) {
+            if (!additional_fields[jQuery( this ).prop( 'name' )]) {
+              additional_fields[jQuery( this ).prop( 'name' )] = {};
+            }
+            additional_fields[jQuery( this ).prop( 'name' )][jQuery( this ).val()] = 1;
+          } else {
+            additional_fields[jQuery( this ).prop( 'name' )] = 1;
           }
-          additional_fields[jQuery( this ).prop( 'name' )][jQuery( this ).val()] = 1;
         }
         else {
-          additional_fields[jQuery( this ).prop( 'name' )] = 1;
+          additional_fields[jQuery( this ).prop( 'name' )] = 0;
         }
       }
       if( 'text' === jQuery(this).attr('type') ) {
@@ -452,7 +463,7 @@ function efml_reset_proxy() {
 /**
  * Start import of single URL.
  *
- * @param url The URLs to import.
+ * @param url The URL to import.
  * @param login The login to use for import.
  * @param password The password to use for import.
  * @param additional_fields Additional fields added by extensions.
@@ -642,8 +653,9 @@ function efml_sync_save_config() {
  * @param login The login (optional).
  * @param password The password (optional).
  * @param api_key The API Key (optional).
+ * @param term_id The used term (optional).
  */
-function efml_save_as_directory( type, url, login, password, api_key ) {
+function efml_save_as_directory( type, url, login, password, api_key, term_id ) {
   jQuery.ajax({
     url: efmlJsVars.ajax_url,
     type: 'POST',
@@ -654,6 +666,7 @@ function efml_save_as_directory( type, url, login, password, api_key ) {
       login: login,
       password: password,
       api_key: api_key,
+      term_id: term_id,
       nonce: efmlJsVars.add_archive_nonce,
     },
     error: function( jqXHR, textStatus, errorThrown ) {