+++ app/ExternalFiles/Files.php
@@ -11,6 +11,7 @@ namespace ExternalFilesInMediaLibrary\ExternalFiles;
 defined( 'ABSPATH' ) || exit;
 
 use easyDirectoryListingForWordPress\Taxonomy;
+use ExternalFilesInMediaLibrary\Dependencies\easySettingsForWordPress\Settings;
 use ExternalFilesInMediaLibrary\Plugin\Admin\Directory_Listing;
 use ExternalFilesInMediaLibrary\Plugin\Helper;
 use ExternalFilesInMediaLibrary\Plugin\Log;
@@ -70,6 +71,9 @@ class Files {
 		// initialize the synchronization.
 		Synchronization::get_instance()->init();
 
+		// initialize REST API support.
+		Rest::get_instance()->init();
+
 		// misc.
 		add_action( 'add_meta_boxes_attachment', array( $this, 'add_media_box' ), 20, 1 );
 
@@ -98,6 +102,9 @@ class Files {
 		add_filter( 'eml_http_directory_regex', array( $this, 'use_link_regex' ), 10, 2 );
 		add_filter( 'eml_help_tabs', array( $this, 'add_help' ), 20 );
 		add_filter( 'eml_file_import_attachment', array( $this, 'add_file_date' ), 10, 3 );
+		add_filter( 'eml_import_fields', array( $this, 'add_date_option_in_form' ) );
+		add_filter( 'eml_http_save_local', array( $this, 'import_local_on_real_import' ) );
+		add_filter( 'eml_file_import_attachment', array( $this, 'add_title_on_real_import' ), 10, 3 );
 
 		// add admin actions.
 		add_action( 'admin_action_eml_reset_thumbnails', array( $this, 'reset_thumbnails_by_request' ) );
@@ -542,7 +549,16 @@ class Files {
 		?>
 		<div class="misc-pub-external-file">
 		<p>
-			<?php echo esc_html__( 'External URL of this file:', 'external-files-in-media-library' ); ?><br><a href="<?php echo esc_url( $url ); ?>" title="<?php echo esc_attr( $url ); ?>"><?php echo esc_html( $url_to_show ); ?></a>
+			<?php echo esc_html__( 'External URL of this file:', 'external-files-in-media-library' ); ?><br>
+			<?php
+			if ( ! empty( esc_url( $url ) ) ) {
+				?>
+					<a href="<?php echo esc_url( $url ); ?>" title="<?php echo esc_attr( $url ); ?>"><?php echo esc_html( $url_to_show ); ?></a>
+				<?php
+			} else {
+				echo esc_html( $url );
+			}
+			?>
 		</p>
 		</div>
 		<ul class="misc-pub-external-file">
@@ -1338,8 +1354,16 @@ class Files {
 	 * @noinspection PhpUnusedParameterInspection
 	 */
 	public function add_file_date( array $post_array, string $url, array $file_data ): array {
-		// bail if setting is disabled.
-		if ( 1 !== absint( get_option( 'eml_use_file_dates' ) ) ) {
+		// get value from request.
+		$use_date = isset( $_POST['additional_fields']['use_dates'] ) ? absint( $_POST['additional_fields']['use_dates'] ) : -1;
+
+		// bail if not set from request and global setting not enabled.
+		if( -1 === $use_date && 1 !== absint( get_option( 'eml_use_file_dates' ) ) ) {
+			return $post_array;
+		}
+
+		// bail if not enabled in request.
+		if( 0 === $use_date ) {
 			return $post_array;
 		}
 
@@ -1408,6 +1432,19 @@ class Files {
 			$api_key = '';
 		}
 
+		// get the credentials from the used term.
+		$term_id = absint( filter_input( INPUT_POST, 'term_id', FILTER_SANITIZE_NUMBER_INT ) );
+		if ( $term_id > 0 ) {
+			// get the term.
+			$term_data = Taxonomy::get_instance()->get_entry( $term_id );
+
+			if ( ! empty( $term_data ) ) {
+				$login    = $term_data['login'];
+				$password = $term_data['password'];
+				$api_key  = $term_data['api_key'];
+			}
+		}
+
 		// add the archive.
 		Taxonomy::get_instance()->add( $type, $url, $login, $password, $api_key );
 
@@ -1432,5 +1469,60 @@ class Files {
 												),
 			)
 		);
+	}
+
+	/**
+	 * Add a checkbox to mark the fields to add them to queue.
+	 *
+	 * @param array<int,string> $fields List of fields in form.
+	 *
+	 * @return array<int,string>
+	 */
+	public function add_date_option_in_form( array $fields ): array {
+		// add the field to enable queue-upload.
+		$fields[] = '<label for="use_dates"><input type="checkbox" name="use_dates" id="use_dates" value="1" class="eml-use-for-import"' . ( 1 === absint( get_option( 'eml_use_file_dates' ) ) ? ' checked="checked"' : '' ) . '> ' . esc_html__( 'Use external file dates.', 'external-files-in-media-library' ) . ' <a href="' . esc_url( \ExternalFilesInMediaLibrary\Plugin\Settings::get_instance()->get_url( 'eml_advanced' ) ) . '" target="_blank"><span class="dashicons dashicons-admin-generic"></span></a></label>';
+
+		// return the resulting fields.
+		return $fields;
+	}
+
+	/**
+	 * Return true if real import is enabled to force local saving of each file.
+	 *
+	 * @param bool $result The result.
+	 *
+	 * @return bool
+	 */
+	public function import_local_on_real_import( bool $result ): bool {
+		// bail if setting is disabled to use the generated value.
+		if( 1 !== absint( get_option( 'eml_directory_listing_real_import' ) ) ) {
+			return $result;
+		}
+
+		// return true to force local saving of this file.
+		return true;
+	}
+
+	/**
+	 * Add title for file if real import is enabled.
+	 *
+	 * @param array<string,mixed> $post_array The attachment settings.
+	 * @param string              $url        The requested external URL.
+	 * @param array<string,mixed> $file_data  List of file settings detected by importer.
+	 *
+	 * @return array<string,mixed>
+	 * @noinspection PhpUnusedParameterInspection
+	 */
+	public function add_title_on_real_import( array $post_array, string $url, array $file_data ): array {
+		// bail if setting is disabled to use the generated value.
+		if( 1 !== absint( get_option( 'eml_directory_listing_real_import' ) ) ) {
+			return $post_array;
+		}
+
+		// add the title.
+		$post_array['post_title'] = $file_data['title'];
+
+		// return the resulting array.
+		return $post_array;
 	}
 }